+++ learningrecord.h
@@ -9,17 +9,18 @@ class LearningRecord
 {
 public:
     LearningRecord(const QString& iFile, const QString& dFile);
-    void addError(const std::vector<std::pair<double, double>>& errors,
-                  const std::vector<double>& activation);
-    std::vector<double> returnDelta() const;
-    std::vector<double> returnError() const;
-    std::vector<double> returnActivation() const;
+    void addError(const std::vector<std::pair<long double,
+                  long double>>& errors,
+                  const std::vector<long double>& activation);
+    std::vector<long double> returnDelta() const;
+    std::vector<long double> returnError() const;
+    std::vector<long double> returnActivation() const;
 
 private:
     const QString imageFile;
     const QString dataFile;
-    std::vector<std::pair<double, double>> errRecord;
-    std::vector<double> active;
+    std::vector<std::pair<long double, long double>> errRecord;
+    std::vector<long double> active;
 };
 
 #endif // LEARNINGRECORD_H
